refactoring views to custom views
